refactor: split ai_services.py into services/ai/ (8 files)
Phase 2: AI service split
- Extract infra (RateLimiter, CircuitBreaker, LRUCache, PersistentCache) to services/ai/infra/
- Extract base service class to services/ai/base.py
- Split LLM/Vision/TTS/ASR services to individual modules
- Create AIServiceManager in services/ai/manager.py
- Make ai_services.py a thin compat layer with deprecation warning
- Update service_manager.py to use new paths (break circular import)
